+++ crates/re_viewer/src/ui/mod.rs
@@ -1,6 +1,4 @@
-mod blueprint;
-mod blueprint_load;
-mod blueprint_sync;
+mod blueprint_panel;
 mod mobile_warning_ui;
 mod rerun_menu;
 mod selection_history_ui;
@@ -10,10 +8,10 @@ mod wait_screen_ui;
 pub(crate) mod memory_panel;
 pub(crate) mod selection_panel;
 
+pub use blueprint_panel::show_blueprint_panel;
 // ----
 
 pub(crate) use {
-    self::blueprint::Blueprint, self::mobile_warning_ui::mobile_warning_ui,
-    self::rerun_menu::rerun_menu_button_ui, self::top_panel::top_panel,
-    self::wait_screen_ui::wait_screen_ui,
+    self::mobile_warning_ui::mobile_warning_ui, self::rerun_menu::rerun_menu_button_ui,
+    self::top_panel::top_panel, self::wait_screen_ui::wait_screen_ui,
 };
