+++ crates/re_viewer/src/ui/selection_panel.rs
@@ -5,9 +5,7 @@ use re_data_store::{ColorMapper, Colormap, EditableAutoValue, EntityPath, Entity
 use re_data_ui::{item_ui, DataUi};
 use re_log_types::TimeType;
 use re_viewer_context::{Item, SpaceViewId, UiVerbosity, ViewerContext};
-use re_viewport::{Viewport, ViewportState};
-
-use crate::ui::Blueprint;
+use re_viewport::{Viewport, ViewportBlueprint, ViewportState};
 
 use super::selection_history_ui::SelectionHistoryUi;
 
@@ -26,7 +24,7 @@ impl SelectionPanel {
         viewport_state: &mut ViewportState,
         ctx: &mut ViewerContext<'_>,
         ui: &mut egui::Ui,
-        blueprint: &mut Blueprint<'_>,
+        blueprint: &mut ViewportBlueprint<'_>,
         expanded: bool,
     ) {
         let screen_width = ui.ctx().screen_rect().width();
@@ -80,7 +78,7 @@ impl SelectionPanel {
         viewport_state: &mut ViewportState,
         ctx: &mut ViewerContext<'_>,
         ui: &mut egui::Ui,
-        blueprint: &mut Blueprint<'_>,
+        blueprint: &mut ViewportBlueprint<'_>,
     ) {
         re_tracing::profile_function!();
 
@@ -216,7 +214,7 @@ fn blueprint_ui(
     viewport_state: &mut ViewportState,
     ui: &mut egui::Ui,
     ctx: &mut ViewerContext<'_>,
-    blueprint: &mut Blueprint<'_>,
+    blueprint: &mut ViewportBlueprint<'_>,
     item: &Item,
 ) {
     match item {
@@ -307,7 +305,7 @@ fn list_existing_data_blueprints(
     ui: &mut egui::Ui,
     ctx: &mut ViewerContext<'_>,
     entity_path: &EntityPath,
-    blueprint: &Blueprint<'_>,
+    blueprint: &ViewportBlueprint<'_>,
 ) {
     let space_views_with_path = blueprint
         .viewport
